Initialize the repository with Devember code
During December 2017 I rewrote my hyperhelp project from the ground
up. As a part of that project, I also developed the code here as the
first steps in a package for providing assistance in writing help
files for hyperhelp.

This commit represents the code as of the end of the last day of
Devember, squashed into a single commit.