Get prepared for multi-package help.
Changes are being made in the core of HyperHelp to allow a single
package to contain more than one help index, allowing for the help of
multiple packages to be contained in a single package for
distribution.

As a part of that change, the help index has a new required key that
specifies the package that the index is intended for, so it has been
added here.

As yet this is not a feature of hyperhelp, so this makes the index
essentially broken unless you also have the newer version of the code,
although these commits will not be made public until all packages have
been updated accordingly.