Apply fixes for new package sources
Various commands in the authoring tools worked under the assumption
that they could determine the package that an opened file was for
based on it's filename and relying on it being stored somewhere in the
Packages folder.

This is no longer possible because the package for a help system is
now a configuration element instead of being based directly on the
package the file is stored in.

To get around this, the code now looks up the appropriate package
based on its document root, which allows the linting code and some
other commands to work.

It has been noted as well that there is likely some inherent interplay
problems with the authoring tools and their use of filenames in that
opened files use the file system convention of the platform but
package resources are always in Posix format.

As such, these will need to be revisited at a later date.