Enhance linting to work with new anchors/links
A recent change to hyperhelp and how links and anchors work allow for
them to have different topics than their visible text and (in the case
of links) to link to topics in other packages.

The linter was still naively using the text of the link/anchor to do the
lint, so this has been modified.

In addition there are now some new linting rules in place to catch
things like links to missing packages, malformed links, and links to
missing package files.

Most of these are also handled internally in hyperhelp in the link hover
popup to provide more contextual errors, so it seems like a good idea to
also support them for linting as well.