Include context menus for easier editing
This includes context menu commands for directly editing either the
help file currently being viewed OR the help index for the current file.

In the first case, this allows you to edit the source of any viewed file
by using the context menu in the help file.

In the second case, any open help file (including the help view) which
is contained in the document root of a loaded package is eligible for
editing the index that file is contained in.