Change bindings and linting for new syntax rules
Recent commits to the HyperHelp syntax and core make it now possible to
have white space in anchors, links and key bindings where previously it
was not possible.

As a result, the core was cleaned up to no longer do manipulations to
convert spaces to and from nonbreaking spaces.

Consequently we have now modified the authoring tools so that there is
no longer a binding on space to insert that character in these cases as
regular space is now valid instead.

Additionally the linter has grown a couple of extra checks to tell you
if a link or anchor contains these characters, to allow for cleanup of
older files.