Emergency fix to the linter to support windows
While creating the video for HyperHelpAuthor, it was noted that the
linter is broken because it infers paths as if they have posix file path
separators, which is not the case on windows at all.

As an expedient fix, this swaps those for the appropriate type should
they exist. There is a better fix for this (such as using posixpath
instead), but this gets things working without a lot of other things
potentially breaking.