Move commands into the source directory
As the last phase of our refactor, this takes the super scary action of
splitting the single `commands.py` file into separate files for each
command, and settings things up so that they are properly conveyed back
to Sublime so that they will operate as expected.

This last step makes the code much more readable, though we reserve the
right for future refactors to do the same thing as this for the linters
as well.