Split the linter code into a separate sub-module
In order to make linting easier to follow, the single linter file has 
been split up into a separate sub-module, much like was done previously
for the commands.

The `linter` module contains only the linters, with the support code for
the linters being in separate files now in the `src` directory.

This makes the linting code a lot easier to understand overall.