Check for files that are not linked by other files
Add a new linter which checks for files that appear in the help index
but not in links in any other help file. This is an indication that it's
possible that the user can't view this help because there is no way to
navigate there from a link.

This is somewhat of an overlap on the recently added linter that checks
for files not in the table of contents; if both linters generate a
warning the file is definitely not linked.

For our purposes here, links that are broken, to files outside of the
current package, or which are linking a file to itself are ignored for
this check. In addition we always assume that `index.txt` is a linked
file because it's the default topic in help navigation if no other topic
is provided.