Send lint output to a view when enabled
When the appropriate setting is enabled (it is currently disabled by
default), send the output of the lint to a view in the current window
instead of to a panel.

Most of this functionality revolves around finding or creating the view
to use in the window (we want to re-use where possible) since the lint
output already has the ability to be displayed in any view you happen
to provide. 

That said, a small amount of refactoring was done here to support this
since we need new information at the call site in order to pull this
off.