Fix invalid link linting
Although the linter is smart enough to be able to detect an invalid link
and report the error, it's not smart enough to test if the link is
invalid before testing to see if it contains undesirable characters.

This reorders some checks so that we only test the content of the link
topic for being invalid if we actually got a topic.